+++ FBSDKGamingServicesKit/LegacyGamingServicesTests/Helpers/SampleURLs.swift
@@ -16,23 +16,9 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-@import LegacyGamingServices;
-@import XCTest;
+import Foundation
 
-#import "FBSDKApplicationDelegate+Testing.h"
-#import "FBSDKCoreKit+Internal.h"
-#import "FBSDKFileHandleCreating.h"
-#import "FBSDKFileHandleFactory.h"
-#import "FBSDKFileHandling.h"
-#import "FBSDKFriendFinderDialog+Testing.h"
-#import "FBSDKGamingImageUploader+Testing.h"
-#import "FBSDKGamingPayloadObserver+Testing.h"
-#import "FBSDKGamingServiceController.h"
-#import "FBSDKGamingServiceControllerCreating.h"
-#import "FBSDKGamingServiceControllerFactory.h"
-#import "FBSDKGamingServiceControllerProtocol.h"
-#import "FBSDKGamingVideoUploader+Testing.h"
-#import "FBSDKVideoUploader.h"
-#import "FBSDKVideoUploaderCreating.h"
-#import "FBSDKVideoUploaderFactory.h"
-#import "FBSDKVideoUploading.h"
+// swiftlint:disable force_unwrapping
+enum SampleUrls {
+  static let valid = URL(string: "https://www.example.com")!
+}